backend changes: added a delete endpoint, added filters to the search changed the database structure to closer align with the scope
frontend changes:
added a delete button in the admin page
made view articles and search the same page
added filters for SEPractice, perspective, and year the book was published
changed the submit article page to work with the changes made to the database structure in the backend.
turned the popup article display into a module that can be used anywhere, added the popup display to the sortable table so that it is always availble with the sortabletable
added functionality to sort by any column; ascending or descending.